requirements.txt
# requirements.txt
plotly>=6.0
PyYAML>=6.0
dash>=2.17
pandas>=2.0
requests>=2.31